+++ crates/client/sync/src/l2.rs
@@ -150,8 +150,6 @@ pub async fn sync<C>(
     // Have 10 fetches in parallel at once, using futures Buffered
     let fetch_stream = stream::iter(fetch_stream).buffered(10);
     let (fetch_stream_sender, mut fetch_stream_receiver) = mpsc::channel(10);
-    let (state_update_sender, mut state_update_receiver) = mpsc::channel(10);
-    let (class_update_sender, mut class_update_receiver) = mpsc::channel(10);
 
     tokio::select!(
         // update highest block hash and number, update pending block and state update
@@ -230,25 +228,21 @@ pub async fn sync<C>(
                         block_sender.send(block_conv).await.expect("block reciever channel is closed");
                     },
                     async {
-                        // Now send state_update, which moves it. This will be received
-                        // by QueryBlockConsensusDataProvider in deoxys/crates/node/src/service.rs
-                        state_update_sender
-                            .send((block_n, state_update))
-                            .await
-                            .expect("state updater is not running");
+                        if store_state_update(block_n, state_update).await.is_err() {
+                            log::info!("❗ Failed to store state update for block {block_n}");
+                        };
                     },
                     async {
-                        // do the same to class update
-                        class_update_sender
-                            .send((block_n, ClassUpdateWrapper(class_update)))
-                            .await
-                            .expect("class updater is not running");
+                        if store_class_update(block_n, ClassUpdateWrapper(class_update)).await.is_err() {
+                            log::info!("❗ Failed to store class update for block {block_n}");
+                        };
+                    },
+                    async {
+                        let start = std::time::Instant::now();
+                        create_block(&mut command_sink, &mut last_block_hash).await.expect("creating block");
+                        log::debug!("end create_block: {:?}", std::time::Instant::now() - start);
                     }
                 );
-
-                let start = std::time::Instant::now();
-                create_block(&mut command_sink, &mut last_block_hash).await.expect("creating block");
-                log::debug!("end create_block: {:?}", std::time::Instant::now() - start);
                 block_n += 1;
 
                 // compact DB every 1k blocks
@@ -257,22 +251,6 @@ pub async fn sync<C>(
                 }
             }
         } => {},
-        // store state updates
-        _ = async {
-            while let Some((block_number, state_update)) = pin!(state_update_receiver.recv()).await {
-                if store_state_update(block_number, state_update).await.is_err() {
-                    log::info!("❗ Failed to store state update for block {block_number}");
-                };
-            }
-        } => {},
-        // store class udpate
-        _ = async {
-            while let Some((block_number, class_update)) = pin!(class_update_receiver.recv()).await {
-                if store_class_update(block_number, class_update).await.is_err() {
-                    log::info!("❗ Failed to store class update for block {block_number}");
-                };
-            }
-        } => {}
     );
 
     log::debug!("L2 sync finished :)");