+++ CHANGELOG.md
@@ -2,6 +2,8 @@ git # Deoxys Changelog
 
 ## Next release
 
+- fix: class and store updates and block desync after ctrl+c
+- fix: compile without libm
 - fix: genesis state_update
 - refactor: optimize get_class_at
 - fix: crash build genesis on restart